Move HTTP poet harness to its own file
