+++ activation/poet_test.go
@@ -47,7 +47,7 @@ func TestHTTPPoet(t *testing.T) {
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
-	c, err := activation.NewHTTPPoetHarness(ctx, poetDir, activation.WithGateway(gtw.Target()))
+	c, err := activation.NewHTTPPoetTestHarness(ctx, poetDir, activation.WithGateway(gtw.Target()))
 	r.NoError(err)
 	r.NotNil(c)
